# stocksense-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockSense API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "List categories",
        "operationId": "list_categories_api_v1_categories_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CategoryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create category",
        "operationId": "create_category_api_v1_categories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{category_id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category",
        "operationId": "get_category_api_v1_categories__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Categories"
        ],
        "summary": "Update category",
        "operationId": "update_category_api_v1_categories__category_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/adjustments": {
      "get": {
        "tags": [
          "Adjustments"
        ],
        "summary": "List Adjustments",
        "operationId": "list_adjustments_api_v1_adjustments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdjustmentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Adjustments"
        ],
        "summary": "Create Adjustment",
        "operationId": "create_adjustment_api_v1_adjustments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adjustments/{adjustment_id}": {
      "get": {
        "tags": [
          "Adjustments"
        ],
        "summary": "Get Adjustment",
        "operationId": "get_adjustment_api_v1_adjustments__adjustment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "adjustment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Adjustment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adjustments/{adjustment_id}/validate": {
      "post": {
        "tags": [
          "Adjustments"
        ],
        "summary": "Validate Adjustment",
        "operationId": "validate_adjustment_api_v1_adjustments__adjustment_id__validate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "adjustment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Adjustment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adjustments/{adjustment_id}/cancel": {
      "post": {
        "tags": [
          "Adjustments"
        ],
        "summary": "Cancel Adjustment",
        "operationId": "cancel_adjustment_api_v1_adjustments__adjustment_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "adjustment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Adjustment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Dashboard",
        "operationId": "get_dashboard_api_v1_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/advanced": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Advanced Dashboard",
        "operationId": "get_advanced_dashboard_api_v1_dashboard_advanced_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancedDashboardSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdjustmentCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "location_id": {
            "type": "string",
            "format": "uuid",
            "title": "Location Id"
          },
          "physical_quantity": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Physical Quantity"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "location_id",
          "physical_quantity",
          "reason"
        ],
        "title": "AdjustmentCreate"
      },
      "AdjustmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "location_id": {
            "type": "string",
            "format": "uuid",
            "title": "Location Id"
          },
          "system_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "System Quantity"
          },
          "physical_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Physical Quantity"
          },
          "difference": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Difference"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reference",
          "product_id",
          "location_id",
          "system_quantity",
          "physical_quantity",
          "difference",
          "reason",
          "status",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "AdjustmentResponse"
      },
      "AdvancedDashboardSummary": {
        "properties": {
          "total_products": {
            "type": "integer",
            "title": "Total Products"
          },
          "active_products": {
            "type": "integer",
            "title": "Active Products"
          },
          "low_stock_products": {
            "type": "integer",
            "title": "Low Stock Products"
          },
          "out_of_stock_products": {
            "type": "integer",
            "title": "Out Of Stock Products"
          },
          "total_on_hand": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total On Hand"
          },
          "total_reserved": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Reserved"
          },
          "total_free_to_use": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Free To Use"
          },
          "total_stock_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Stock Value"
          },
          "pending_receipts": {
            "type": "integer",
            "title": "Pending Receipts"
          },
          "late_receipts": {
            "type": "integer",
            "title": "Late Receipts"
          },
          "waiting_receipts": {
            "type": "integer",
            "title": "Waiting Receipts"
          },
          "pending_deliveries": {
            "type": "integer",
            "title": "Pending Deliveries"
          },
          "late_deliveries": {
            "type": "integer",
            "title": "Late Deliveries"
          },
          "waiting_deliveries": {
            "type": "integer",
            "title": "Waiting Deliveries"
          },
          "scheduled_transfers": {
            "type": "integer",
            "title": "Scheduled Transfers"
          },
          "ready_transfers": {
            "type": "integer",
            "title": "Ready Transfers"
          },
          "warehouses": {
            "items": {
              "$ref": "#/components/schemas/DashboardWarehouseTotal"
            },
            "type": "array",
            "title": "Warehouses"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/DashboardLocationTotal"
            },
            "type": "array",
            "title": "Locations"
          },
          "recent_movements": {
            "items": {
              "$ref": "#/components/schemas/StockLedgerResponse"
            },
            "type": "array",
            "title": "Recent Movements"
          }
        },
        "type": "object",
        "required": [
          "total_products",
          "active_products",
          "low_stock_products",
          "out_of_stock_products",
          "total_on_hand",
          "total_reserved",
          "total_free_to_use",
          "total_stock_value",
          "pending_receipts",
          "late_receipts",
          "waiting_receipts",
          "pending_deliveries",
          "late_deliveries",
          "waiting_deliveries",
          "scheduled_transfers",
          "ready_transfers",
          "warehouses",
          "locations",
          "recent_movements"
        ],
        "title": "AdvancedDashboardSummary"
      },
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate"
      },
      "CategoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "title": "CategoryResponse"
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CategoryUpdate"
      },
...
      "DashboardSummary": {
        "properties": {
          "total_products_in_stock": {
            "type": "integer",
            "title": "Total Products In Stock"
          },
          "low_stock_count": {
            "type": "integer",
            "title": "Low Stock Count"
          },
          "out_of_stock_count": {
            "type": "integer",
            "title": "Out Of Stock Count"
          },
          "pending_receipts": {
            "$ref": "#/components/schemas/ReceiptBreakdown"
          },
          "pending_deliveries": {
            "$ref": "#/components/schemas/DeliveryBreakdown"
          },
          "scheduled_transfers": {
            "type": "integer",
            "title": "Scheduled Transfers"
          }
        },
        "type": "object",
        "required": [
          "total_products_in_stock",
          "low_stock_count",
          "out_of_stock_count",
          "pending_receipts",
          "pending_deliveries",
          "scheduled_transfers"
        ],
        "title": "DashboardSummary"
      },
...
      "DeliveryBreakdown": {
        "properties": {
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "late": {
            "type": "integer",
            "title": "Late"
          },
          "waiting": {
            "type": "integer",
            "title": "Waiting"
          },
          "to_deliver": {
            "type": "integer",
            "title": "To Deliver"
          }
        },
        "type": "object",
        "required": [
          "pending",
          "late",
          "waiting",
          "to_deliver"
        ],
        "title": "DeliveryBreakdown"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "WAITING",
          "READY",
          "DONE",
          "CANCELED"
        ],
        "title": "OperationStatus",
        "description": "Lifecycle states shared by operation documents."
      },
      "Page_AdjustmentResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdjustmentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[AdjustmentResponse]"
      },
      "Page_CategoryResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[CategoryResponse]"
      },
...
      "ReceiptBreakdown": {
        "properties": {
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "late": {
            "type": "integer",
            "title": "Late"
          },
          "waiting": {
            "type": "integer",
            "title": "Waiting"
          },
          "to_receive": {
            "type": "integer",
            "title": "To Receive"
          }
        },
        "type": "object",
        "required": [
          "pending",
          "late",
          "waiting",
          "to_receive"
        ],
        "title": "ReceiptBreakdown"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}